requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.2

tensorflow-cpu==2.17.0
keras==3.3.3
scipy==1.12.0
python-multipart>=0.0.7
